Trim redundant tests and consolidate checks
