+++ test/full_game_lifecycle_test.dart
@@ -1,5 +1,6 @@
 import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
+import 'package:flame_audio/flame_audio.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/widgets.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -14,8 +15,32 @@ import 'package:space_game/ui/game_over_overlay.dart';
 import 'package:space_game/ui/hud_overlay.dart';
 import 'package:space_game/ui/menu_overlay.dart';
 
+class _FakeAudioCache extends AudioCache {
+  @override
+  Future<Uri> load(String file) async => Uri();
+
+  @override
+  Future<void> clearAll() async {}
+}
+
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
+
+  test('asset loaders can be called multiple times and missing asset throws',
+      () async {
+    Flame.images.clearCache();
+    FlameAudio.audioCache = _FakeAudioCache();
+
+    await Assets.loadEssential();
+    await Assets.loadEssential();
+    await Assets.loadRemaining();
+    await Assets.loadRemaining();
+
+    expect(
+      () => Flame.images.fromCache('does_not_exist.png'),
+      throwsA(isA<AssertionError>()),
+    );
+  });
 
   test('full lifecycle from start to menu', () async {
     SharedPreferences.setMockInitialValues({});
