+++ test/starfield_density_guard_test.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flame/game.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -15,15 +17,24 @@ void main() {
     expect(negativeLayer.debugTileStarRadii(0, 0), isEmpty);
   });
 
-  test('zero density layer does not crash during update', () async {
+  test('zero density layer does not crash during update or debug render',
+      () async {
     final game = FlameGame();
     game.onGameResize(Vector2.all(256));
-    final starfield =
-        StarfieldComponent(layers: const [StarfieldLayerConfig(density: 0)]);
+    final starfield = StarfieldComponent(
+      layers: const [StarfieldLayerConfig(density: 0)],
+      debugDrawTiles: true,
+    );
     await game.add(starfield);
 
     starfield.update(0);
     await starfield.debugWaitForPending();
+
+    final recorder = PictureRecorder();
+    final canvas = Canvas(recorder);
+
+    expect(starfield.debugDrawTiles, isTrue);
+    starfield.render(canvas); // should draw without throwing
 
     expect(starfield.debugTileStarRadii(0, 0), isEmpty);
   });
